Update exercise-28: clarify that \xvec is a vector
Given the dimensions of a matrix A, exercise-28 asks students to think about the dimensions of x and Ax. It's customary for something called x to be a vector, but I think it's worth clarifying to eliminate ambiguity. If x could be a matrix (and a novice linear algebra student might think it could be), then x could have dimensions 2201 \times n, in which case Ax would have dimensions 135 \times n.